Introduce a 'seam' via intermediate 'os' package w/ function variables.
A 'seam', a term coined by Michael Feathers (author of "Working with
Legacy Code"), refers to any mechanism that allows the behavior of
code to change without the code itself needing to change.

In this case, the 'seam' is a very thin wrapper 'os' package that
forwards to the actual standard library 'os' package. The functions are
actually mutable variable references so they can be reassigned with
faked implementations during testing.

This approach, while convenient, is not without downsides. If multiple
concurrent tests depend on this mutable behavior, locks/mutexes will be
required to prevent unpredictable test failures and data races.